Split out replay tests into folder and create test data
