+++ tests/test-replay/cookiedozer_load.json
@@ -1,0 +1,1 @@
+{"cookiecutter": {"version": "0.1.0", "email": "raphael@hackebrot.de", "full_name": "Raphael Pierzina", "github_username": "hackebrot"}}
